projection prototype.

+++ src/screen.c
@@ -1,11 +1,11 @@
 #include "screen.h"
 
-const int SCREEN_WIDTH = 768;
-const int SCREEN_HEIGHT = 432;
+const float SCREEN_WIDTH = 768;
+const float SCREEN_HEIGHT = 432;
 //const int SCREEN_WIDTH = 640;
 //const int SCREEN_HEIGHT = 480;
 
-const float SCREEN_ASPECT_RATIO =  (float) SCREEN_HEIGHT / (float) SCREEN_WIDTH;
+const float SCREEN_ASPECT_RATIO =  SCREEN_HEIGHT / SCREEN_WIDTH;
 const int SCREEN_UNITS_WIDTH = 160;
 const int SCREEN_UNITS_HEIGHT = 90;
